added paths and rename folder client to clients
