+++ src/clients/cardtokens/create/create.spec.ts
@@ -15,7 +15,7 @@ describe('Testing card tokens, create', () => {
 
         await create({body: body, config: client})
 
-        const spyFetch = jest.spyOn(RestClient, 'fetch') 
+        const spyFetch = jest.spyOn(RestClient, 'fetch')
         expect(spyFetch).toHaveBeenCalledWith("/card_tokens", {"headers": {"Authorization": "Bearer token"},
           "timeout": 5000, "body": "{\"card_id\":\"11111111\",\"security_code\":\"123\"}",
           "method": "POST"})
