+++ src/index.ts
@@ -1,6 +1,6 @@
 import { MercadoPagoConfig } from './MercadoPagoConfig';
-import { Payments } from './client/payments';
-import { CardToken } from './client/cardtokens';
+import { Payments } from './clients/payments';
+import { CardToken } from './clients/cardtokens';
 
 
 export default MercadoPagoConfig;
