+++ tsconfig.json
@@ -8,8 +8,14 @@
     "target": "ES2017",
     "sourceMap": false,
     "outDir": "./dist",
-    "baseUrl": "./",
-    "incremental": true
+    "baseUrl": ".",
+    "rootDir": "./src",
+    "incremental": true,
+    "paths": {
+      "@src/*": ["./src/*"],
+      "@utils/*": ["./src/utils/*"],
+      "@clients/*": ["./src/clients/*"]
+    }
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "test", "lib", "**/*spec.ts"]
